move bit_decomposition
