+++ stdlib/src/blackbox_fallbacks/logic_fallbacks.rs
@@ -1,62 +1,10 @@
-use crate::helpers::VariableStore;
 use acir::{
     acir_field::FieldElement,
-    circuit::{directives::Directive, Opcode},
+    circuit::Opcode,
     native_types::{Expression, Witness},
 };
 
-// Generates opcodes and directives to bit decompose the input `gate`
-// Returns the bits and the updated witness counter
-// TODO:Ideally, we return the updated witness counter, or we require the input
-// TODO to be a VariableStore. We are not doing this because we want migration to
-// TODO be less painful
-pub(crate) fn bit_decomposition(
-    gate: Expression,
-    bit_size: u32,
-    mut num_witness: u32,
-) -> (Vec<Opcode>, Vec<Witness>, u32) {
-    let mut new_gates = Vec::new();
-    let mut variables = VariableStore::new(&mut num_witness);
-
-    // First create a witness for each bit
-    let mut bit_vector = Vec::with_capacity(bit_size as usize);
-    for _ in 0..bit_size {
-        bit_vector.push(variables.new_variable())
-    }
-
-    // Next create a directive which computes those bits.
-    new_gates.push(Opcode::Directive(Directive::ToLeRadix {
-        a: gate.clone(),
-        b: bit_vector.clone(),
-        radix: 2,
-    }));
-
-    // Now apply constraints to the bits such that they are the bit decomposition
-    // of the input and each bit is actually a bit
-    let mut binary_exprs = Vec::new();
-    let mut bit_decomp_constraint = gate;
-    let mut two_pow: FieldElement = FieldElement::one();
-    let two = FieldElement::from(2_i128);
-    for &bit in &bit_vector {
-        // Bit constraint to ensure each bit is a zero or one; bit^2 - bit = 0
-        let mut expr = Expression::default();
-        expr.push_multiplication_term(FieldElement::one(), bit, bit);
-        expr.push_addition_term(-FieldElement::one(), bit);
-        binary_exprs.push(Opcode::Arithmetic(expr));
-
-        // Constraint to ensure that the bits are constrained to be a bit decomposition
-        // of the input
-        // ie \sum 2^i * x_i = input
-        bit_decomp_constraint.push_addition_term(-two_pow, bit);
-        two_pow = two * two_pow;
-    }
-
-    new_gates.extend(binary_exprs);
-    bit_decomp_constraint.sort(); // TODO: we have an issue open to check if this is needed. Ideally, we remove it.
-    new_gates.push(Opcode::Arithmetic(bit_decomp_constraint));
-
-    (new_gates, bit_vector, variables.finalize())
-}
+use super::utils::bit_decomposition;
 
 // Range constraint
 pub fn range(gate: Expression, bit_size: u32, num_witness: u32) -> (u32, Vec<Opcode>) {
